# lib/api-zod/src/generated/types/bankAdvice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Milk Bill Reporting System API
 * OpenAPI spec version: 0.1.0
 */
import type { BankAdviceStatus } from "./bankAdviceStatus";

export interface BankAdvice {
  id: number;
  billId: number;
  billNumber: string;
  societyName: string;
  bankName?: string;
  bankAccount?: string;
  bankIfsc?: string;
  amount: number;
  adviceDate: Date;
  status: BankAdviceStatus;
}